Don't underscore-prefix extract_labware_definitions submodule.
Even though this submodule is private, let's keep things consistent with the other submodules for now.
